requirements.txt
torch
torchvision
numpy
pillow
fastapi
uvicorn
python-multipart
scikit-image
transformers
